+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add filters and detail columns to rp stock dc
 - Add DATACENTERS column and --dc filter to rp stock cpus
+- Add filters and detail columns to rp stock gpu
 
 ### Changed